+++ services/system/Accounts/plugin/accounts/Cargo.toml
@@ -6,7 +6,7 @@ edition = "2021"
 [dependencies]
 url = "2.5.2"
 wit-bindgen-rt = { version = "0.28.0", features = ["bitflags"] }
-psibase = { path = "../../../../rust/psibase/" }
+psibase = { path = "../../../../../rust/psibase/" }
 base64 = "0.22"
 serde = "1.0.204"
 serde_json = "1.0.120"
@@ -28,10 +28,10 @@ package = "accounts:plugin"
 world = "impl"
 
 [package.metadata.component.target.dependencies]
-"host:common" = { path = "../../../user/CommonApi/common/packages/wit/host-common.wit" }
-"host:privileged" = { path = "../../../user/CommonApi/common/packages/wit/host-privileged.wit" }
-"transact:plugin" = { path = "../../Transact/plugin/wit/world.wit" }
-"clientdata:plugin" = { path = "../../../user/ClientData/plugin/wit/world.wit" }
+"host:common" = { path = "../../../../user/CommonApi/common/packages/wit/host-common.wit" }
+"host:privileged" = { path = "../../../../user/CommonApi/common/packages/wit/host-privileged.wit" }
+"transact:plugin" = { path = "../../../Transact/plugin/wit/world.wit" }
+"clientdata:plugin" = { path = "../../../../user/ClientData/plugin/wit/world.wit" }
 
 [package.metadata.component.bindings]
 derives = ["serde::Deserialize"]
