+++ services/system/Accounts/plugin/accounts/src/lib.rs
@@ -1,11 +1,11 @@
 #[allow(warnings)]
 mod bindings;
-mod connection_token;
 mod db;
 mod errors;
 mod helpers;
 mod interfaces;
 mod plugin;
+pub mod tokens;
 
 use plugin::AccountsPlugin;
