+++ services/user/Identity/plugin/Cargo.toml
@@ -23,5 +23,5 @@ world = "impl"
 
 [package.metadata.component.target.dependencies]
 "host:common" = { path = "../../CommonApi/common/packages/wit/host-common.wit" }
-"accounts:plugin" = { path = "../../../system/Accounts/plugin/wit/world.wit" }
-"transact:plugin" = { path = "../../../system/Transact/plugin/wit/world.wit"}
+"accounts:plugin" = { path = "../../../system/Accounts/plugin/accounts/wit/world.wit" }
+"transact:plugin" = { path = "../../../system/Transact/plugin/wit/world.wit" }
